@@ -13,18 +13,17 @@ class ExpandableListView extends BoxScrollView {
   final SliverExpandableChildDelegate builder;
 
   ExpandableListView({
-    Key key,
-    @required this.builder,
+    Key? key,
+    required this.builder,
     bool reverse = false,
-    ScrollController controller,
-    bool primary,
-    ScrollPhysics physics,
+    ScrollController? controller,
+    bool? primary,
+    ScrollPhysics? physics,
     bool shrinkWrap = false,
-    EdgeInsetsGeometry padding,
-    double cacheExtent,
+    EdgeInsetsGeometry? padding,
+    double? cacheExtent,
     DragStartBehavior dragStartBehavior = DragStartBehavior.start,
-  })  : assert(builder != null),
-        super(
+  }) : super(
           key: key,
           scrollDirection: Axis.vertical,
           reverse: reverse,
@@ -34,7 +33,7 @@ class ExpandableListView extends BoxScrollView {
           shrinkWrap: shrinkWrap,
           padding: padding,
           cacheExtent: cacheExtent,
-          semanticChildCount: builder.sectionList?.length ?? 0,
+          semanticChildCount: builder.sectionList.length,
           dragStartBehavior: dragStartBehavior,
         );
 
